multi sshkey module, currently bugged
